통신 CRUD 완료

+++ lib/views/product/list/product_list_view.dart
@@ -1,7 +1,5 @@
 import 'package:data_app/controller/product_controller.dart';
 import 'package:data_app/domain/product/product.dart';
-import 'package:data_app/domain/product/product_http_repository.dart';
-import 'package:data_app/views/components/my_alert_dialog.dart';
 import 'package:data_app/views/product/list/product_list_view_store.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -14,36 +12,42 @@ class ProductListView extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final pm = ref.watch(productListViewStore); // view모델을 바라봄
     final pc = ref.read(productController);
-    final pl = ref.listen(productHttpRepository, (previous, next) {}); // 리빌드 X
+
+    // 빌드 내부에 메서드 호출 사용X
 
     return Scaffold(
       floatingActionButton: FloatingActionButton(
         child: Icon(Icons.add),
         onPressed: () {
-          pc.insert(Product(4, '호박', 20000));
-          showCupertinoDialog(
-            context: context,
-            builder: (context) => MyAlertDialog(msg: "생성 완료"),
-          );
+          pc.insert(Product(id: 0, name: "복숭아", price: 1200));
         },
       ),
       appBar: AppBar(title: Text("product_list_page")),
-      body: ListView.builder(
+      body: _buildListView(pm, pc),
+    );
+  }
+
+  Widget _buildListView(List<Product> pm, ProductController pc) {
+    if (!(pm.length > 0)) {
+      return Center(
+        child: Image.asset(
+          "assets/image/loading.gif",
+          width: 100,
+          height: 100,
+        ),
+      );
+    } else {
+      return ListView.builder(
         itemCount: pm.length,
         itemBuilder: (context, index) => ListTile(
           // key = id값
           key: ValueKey(pm[index].id),
           onTap: () {
             pc.deleteById(pm[index].id);
-            // if (pm[index].id == 4) {
-            //   showCupertinoDialog(
-            //     context: context,
-            //     builder: (context) => MyAlertDialog(msg: "삭제 실패"),
-            //   );
-            // }
           },
           onLongPress: () {
-            pc.updateById(pm[index].id, Product(99, 'test', 60000));
+            pc.updateById(
+                pm[index].id, Product(id: 99, name: '고구마', price: 9999));
           },
           leading: Icon(Icons.wallet_travel),
           title: Text(
@@ -52,7 +56,7 @@ class ProductListView extends ConsumerWidget {
           ),
           subtitle: Text("${pm[index].price}"),
         ),
-      ),
-    );
+      );
+    }
   }
 }
